requirements.txt
python-telegram-bot==20.4
paramiko==3.3.0
cryptography>=39.0.0